update: 更新项目 1.修改部分文件目录 2.添加其他配置支持
