+++ config/config.ini
@@ -2,7 +2,7 @@
 name = Mys_Exch_Goods
 env = local
 branch = main
-version = 2.0.2
+version = 2.0.3
 
 [user_info]
 cookie =
