+++ docs/ABOUT.md
@@ -1,3 +1,6 @@
 ## Mys Exchange Goods
 
 米游社商品兑换工具，心仪的商品仅离你一步之遥~
+项目名称: Mys Exchange Goods
+项目作者: GOOD-AN
+项目地址: https://github.com/GOOD-AN/Mys-Exchange-Goods/
