+++ docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # 更新日志
 
+## [2.0.3.221116 Alpha] - 2022-11-16
+
+### Added
+- 添加了对其他配置项的支持
+- 添加欢迎语
+
+### Changed
+- 部分文件目录重构
+
 ## [2.0.2.221112 Alpha] - 2022-11-12
 
 ### Fixed
