+++ plugin/__init__.py
@@ -1,0 +1,6 @@
+"""
+初始化
+"""
+from .get_info import info_main
+from .exchange_gift import gift_main
+from .other_config import config_main
