+++ tools/__init__.py
@@ -1,5 +1,5 @@
 """
 初始化
 """
-from .com_tool import check_plat, load_config, get_cookie_str, write_config_file, check_update, get_time, md5_encode
+from .com_tool import get_cookie_str, write_config_file, check_update, get_time, md5_encode, init_config
 from .mi_tool import update_cookie, check_cookie, get_gift_detail, check_game_roles, get_point, GAME_NAME, YS_SERVER
